Changes in Belos (tpetra tests) and Tpetra (source and tests) to handle compile warnings: * added some static_cast to handle comparison of signed (int) and unsigned (size_t) types * removed some unused variables
Also, removed commented-out and uneeded bit in Tpetra top-level CMakeLists.txt

Bug #4609:
* Moved implementation of non-templated Tpetra::Distributor methods to a CPP file to be compiled into libtpetra.a




=============================
Automated status information
=============================


Wed Oct 14 08:23:48 EDT 2009



-------------------
Summary of Results:
-------------------


MPI_DEBUG Results:
------------------------

  passed : Trilinos/MPI_DEBUG: passed=11,notpassed=0

  Wed Oct 14 07:50:41 EDT 2009

  Enabled Packages: Belos, Tpetra
  Hostname: zan.ornl.gov
  Source Dir: /Users/cgbaker/Trilinos10
  Build Dir: /Users/cgbaker/Trilinos10/CHECKIN/MPI_DEBUG

  CMake Cache Varibles: -DTPL_ENABLE_MPI:BOOL=ON -DTrilinos_ENABLE_TESTS:BOOL=ON -DCMAKE_BUILD_TYPE:STRING=RELEASE -DTrilinos_ENABLE_DEBUG:BOOL=ON -DTrilinos_ENABLE_CHECKED_STL:BOOL=ON -DTrilinos_ENABLE_EXPLICIT_INSTANTIATION:BOOL=ON -DBUILD_SHARED_LIBS:BOOL=ON -DTrilinos_ENABLE_Fortran:BOOL=OFF -DTrilinos_ENABLE_Zoltan:BOOL=OFF -DTrilinos_ENABLE_Isorropia:BOOL=OFF -DTrilinos_ENABLE_AztecOO:BOOL=OFF -DTrilinos_ENABLE_Galeri:BOOL=OFF -DTrilinos_ENABLE_Amesos:BOOL=OFF -DTrilinos_ENABLE_Ifpack:BOOL=OFF -DTrilinos_ENABLE_ML:BOOL=OFF -DTrilinos_ENABLE_Thyra:BOOL=OFF -DTrilinos_ENABLE_NOX:BOOL=OFF -DTrilinos_ENABLE_GlobiPack:BOOL=OFF -DTrilinos_ENABLE_RTOp:BOOL=OFF -DTrilinos_ENABLE_Shards:BOOL=OFF -DTrilinos_ENABLE_Pamgen:BOOL=OFF -DTrilinos_ENABLE_FEI:BOOL=OFF -DTrilinos_ENABLE_Sacado:BOOL=OFF -DTrilinos_ENABLE_Intrepid:BOOL=OFF -DTrilinos_ENABLE_MOOCHO:BOOL=OFF -DTrilinos_ENABLE_Epetra:BOOL=OFF -DTrilinos_ENABLE_Anasazi:BOOL=OFF -DTrilinos_ENABLE_Belos:BOOL=OFF -DMPI_BASE_DIR:PATH=/opt/openmpi-1.3.3-gcc-4.0 -DTrilinos_ENABLE_Belos:BOOL=ON -DTrilinos_ENABLE_Tpetra:BOOL=ON -DTrilinos_ENABLE_ALL_OPTIONAL_PACKAGES:BOOL=ON -DTrilinos_ENABLE_ALL_FORWARD_DEP_PACKAGES:BOOL=ON

  Make Options: -j2

  CTest Options: -j2

  Update: Not Performed
  Configure: Passed (0.478365067641 min)
  Build: Passed (0.0668979326884 min)
  Test: Passed (0.262223235766 min)

  Test summary:
  -------------

  100% tests passed, 0 tests failed out of 11

  Label Time Summary:
  Belos     =   3.00 sec
  Tpetra    =  23.00 sec

  Total Test time (real) =  15.00 sec


  Final:
  ------

  Total time for MPI_DEBUG = 0.809951333205 minutes


SERIAL_RELEASE Results:
------------------------

  passed : Trilinos/SERIAL_RELEASE: passed=12,notpassed=0

  Wed Oct 14 08:07:44 EDT 2009

  Enabled Packages: Belos, Tpetra
  Hostname: zan.ornl.gov
  Source Dir: /Users/cgbaker/Trilinos10
  Build Dir: /Users/cgbaker/Trilinos10/CHECKIN/SERIAL_RELEASE

  CMake Cache Varibles: -DTrilinos_ENABLE_TESTS:BOOL=ON -DCMAKE_BUILD_TYPE:STRING=RELEASE -DTrilinos_ENABLE_DEBUG:BOOL=OFF -DTrilinos_ENABLE_CHECKED_STL:BOOL=OFF -DTrilinos_ENABLE_EXPLICIT_INSTANTIATION:BOOL=OFF -DBUILD_SHARED_LIBS:BOOL=ON -DTrilinos_ENABLE_Fortran:BOOL=OFF -DTrilinos_ENABLE_Zoltan:BOOL=OFF -DTrilinos_ENABLE_Isorropia:BOOL=OFF -DTrilinos_ENABLE_AztecOO:BOOL=OFF -DTrilinos_ENABLE_Galeri:BOOL=OFF -DTrilinos_ENABLE_Amesos:BOOL=OFF -DTrilinos_ENABLE_Ifpack:BOOL=OFF -DTrilinos_ENABLE_ML:BOOL=OFF -DTrilinos_ENABLE_Thyra:BOOL=OFF -DTrilinos_ENABLE_NOX:BOOL=OFF -DTrilinos_ENABLE_GlobiPack:BOOL=OFF -DTrilinos_ENABLE_RTOp:BOOL=OFF -DTrilinos_ENABLE_Shards:BOOL=OFF -DTrilinos_ENABLE_Pamgen:BOOL=OFF -DTrilinos_ENABLE_FEI:BOOL=OFF -DTrilinos_ENABLE_Sacado:BOOL=OFF -DTrilinos_ENABLE_Intrepid:BOOL=OFF -DTrilinos_ENABLE_MOOCHO:BOOL=OFF -DTrilinos_ENABLE_Epetra:BOOL=OFF -DTrilinos_ENABLE_Anasazi:BOOL=OFF -DTrilinos_ENABLE_Belos:BOOL=OFF -DCMAKE_C_COMPILER=gcc-4.0 -DCMAKE_CXX_COMPILER=g++-4.0 -DTrilinos_ENABLE_Belos:BOOL=ON -DTrilinos_ENABLE_Tpetra:BOOL=ON -DTrilinos_ENABLE_ALL_OPTIONAL_PACKAGES:BOOL=ON -DTrilinos_ENABLE_ALL_FORWARD_DEP_PACKAGES:BOOL=ON

  Make Options: -j2

  CTest Options: -j2

  Update: Not Performed
  Configure: Passed (0.399560681979 min)
  Build: Passed (16.5719685634 min)
  Test: Passed (0.0731598337491 min)

  Test summary:
  -------------

  100% tests passed, 0 tests failed out of 12

  Label Time Summary:
  Belos     =   0.00 sec
  Tpetra    =   6.00 sec

  Total Test time (real) =   3.00 sec


  Final:
  ------

  Total time for SERIAL_RELEASE = 17.0494975487 minutes

+++ packages/belos/tpetra/test/BlockCG/test_bl_cg_complex_hb.cpp
@@ -146,7 +146,7 @@ int main(int argc, char *argv[]) {
     const double *curdval = dvals;
     // set up pointers
     newoffs[0] = 0;
-    for (size_t row=1; row != dim+1; ++row) {
+    for (size_t row=1; row != static_cast<size_t>(dim+1); ++row) {
       newoffs[row] = newoffs[row-1] + rnnz[row-1];
     }
     // reorganize data from column oriented to row oriented, duplicating symmetric part as well
