+++ packages/belos/tpetra/test/BlockCG/test_bl_cg_hb.cpp
@@ -128,10 +128,10 @@ int main(int argc, char *argv[]) {
     // Skip the diagonals, we'll catch them below on the column pass.
     // Remember, the file uses one-based indexing. We'll convert it to zero-based later.
     int curcol_0 = 0, currow_0;
-    for (size_t curnnz_1=1; curnnz_1 <= nnz; ++curnnz_1) {
+    for (size_t curnnz_1=1; curnnz_1 <= static_cast<size_t>(nnz); ++curnnz_1) {
       // if colptr[c] <= curnnz_1 < colptr[c+1], then curnnz_1 belongs to column c
       // invariant: curcol_0 is the column for curnnz_1, i.e., curcol_0 is smallest number such that colptr[curcol_0+1] > curnnz_1
-      while (colptr[curcol_0+1] <= curnnz_1) ++curcol_0;
+      while (static_cast<size_t>(colptr[curcol_0+1]) <= curnnz_1) ++curcol_0;
       // entry curnnz_1 corresponds to (curcol_0, rowind[curnnz_1]) and (rowind[curnnz_1], curcol_0)
       // make sure not to count it twice
       ++rnnz[curcol_0];
@@ -149,7 +149,7 @@ int main(int argc, char *argv[]) {
     int *newinds    = new int[totalnnz];
     // set up pointers
     newoffs[0] = 0;
-    for (size_t row=1; row != dim+1; ++row) {
+    for (size_t row=1; row != static_cast<size_t>(dim+1); ++row) {
       newoffs[row] = newoffs[row-1] + rnnz[row-1];
     }
     // reorganize data from column oriented to row oriented, duplicating symmetric part as well
@@ -157,10 +157,10 @@ int main(int argc, char *argv[]) {
     // use nnz as the number of entries add per row thus far
     std::fill( rnnz.begin(), rnnz.end(), 0 );
     curcol_0 = 0;
-    for (size_t curnnz_1=1; curnnz_1 <= nnz; ++curnnz_1) {
+    for (size_t curnnz_1=1; curnnz_1 <= static_cast<size_t>(nnz); ++curnnz_1) {
       // if colptr[c] <= curnnz_1 < colptr[c+1], then curnnz_1 belongs to column c
       // invariant: curcol_0 is the column for curnnz_1, i.e., curcol_0 is smallest number such that colptr[curcol_0+1] > curnnz_1
-      while (colptr[curcol_0+1] <= curnnz_1) ++curcol_0;
+      while (static_cast<size_t>(colptr[curcol_0+1]) <= curnnz_1) ++curcol_0;
       // entry curnnz_1 corresponds to (curcol_0, rowind[curnnz_1]) and (rowind[curnnz_1], curcol_0)
       // it must be added twice if curcol_0 != rowind[curnnz_1]
       currow_0 = rowind[curnnz_1-1] - 1;
@@ -208,7 +208,7 @@ int main(int argc, char *argv[]) {
   RCP<const Map<int> > map = rcp(new Map<int>(dim,0,comm) );
   RCP<CrsMatrix<ST,int> > A = rcp(new CrsMatrix<ST,int>(map,rnnzmax));
   if (MyPID == 0) {
-    for (size_t row=0; row < dim; ++row) {
+    for (size_t row=0; row < static_cast<size_t>(dim); ++row) {
       const size_t nE = offsets[row+1] - offsets[row];
       if (nE > 0) {
         // add row to matrix
